SpecificGroundComponent
Basic Functionality for the 3x3 Concrete Requirement for Wind Turbines. Component is expandable if we want to use similar things on other objects, but currently just checks for a 3x3 grid of the given Blocktype under the "center" (the block where you place it) of the worldobject. Im not sure about performance, that needs testing, but im already only checking every 100 Ticks