Added diamond
Diamond cut, Raw Diamond. We need to take a look at the Diamond Cut recipe and decide who should do it.

Remove unitypackage because it now contains source code scripts which we're not allowed to distribute. Also changed the changelog, not sure about all the updates you've made Yamor so maybe have a look at that? :)